Tweaks from sonar analysis

+++ odin-core/src/main/java/com/purplepip/odin/music/logic/PatternLogic.java
@@ -6,10 +6,8 @@ import com.purplepip.odin.sequence.DefaultEvent;
 import com.purplepip.odin.sequence.Event;
 import com.purplepip.odin.sequence.MutableTock;
 import com.purplepip.odin.sequence.ScanForwardEvent;
-import com.purplepip.odin.sequence.Sequence;
 import com.purplepip.odin.sequence.Tock;
 import com.purplepip.odin.sequence.logic.AbstractLogic;
-import com.purplepip.odin.sequence.logic.Logic;
 import com.purplepip.odin.sequence.measure.MeasureProvider;
 
 import org.slf4j.Logger;
@@ -33,7 +31,7 @@ public class PatternLogic extends AbstractLogic<Pattern, Note> {
     MutableTock mutableTock = new MutableTock(tock);
     Event<Note> nextEvent;
     boolean on = false;
-    long maxForwardScan = 2 * measureProvider.getBeatsInThisMeasure(mutableTock);
+    int maxForwardScan = 2 * measureProvider.getBeatsInThisMeasure(mutableTock);
     int i = 0;
     while (!on && i < maxForwardScan) {
       mutableTock.increment();
